backend/pyproject.toml
[build-system]
requires = ["setuptools >= 77.0.3"]
build-backend = "setuptools.build_meta"

[project]
name = "fake-bank-backend"
version = "0.1.0"
requires-python = ">=3.12"
authors = [
    {name = "Tanuj Guha Thakurta", email = "tanuj414junat@gmail.com"}
]
dependencies = [
    "alembic>=1.17.1",
    "bcrypt==4.0.1",
    "email-validator>=2.3.0",
    "fastapi[standard]>=0.120.4",
    "passlib[bcrypt]>=1.7.4",
    "psycopg2>=2.9.11",
    "pydantic-settings>=2.11.0",
    "sqlalchemy>=2.0.44",
]

[project.optional-dependencies]
test = [
    "pytest>=9.0.0",
]
dev = [
    "black>=25.9.0",
    "isort>=7.0.0",
    "mypy>=1.18.2",
    "pylint>=4.0.2",
    "ruff>=0.14.4",
]

[tool.setuptools.packages.find]
where = ["."]
